Have fpgrowth output a dict of {itemset->support}, use that for rule generation instead of an InvertedIndex.
When we save an itemset in FPGrowth, we know it's support. So we can
save that too. That means we can use the support during rule generation
instead of calculating it from the InvertedIndex. This can save a lot
of memory as we don't need to load the InvertedIndex. It also can save
a significant amount of time, as calculating the support using the
InvertedIndex can be very slow.